external links
docusauris uses _to_ for internal links and _href_ for external links

This way the link gets the icon to indicate that the link opens in a new tab etc.

In the footer the two community links were done this way but the More and Legal Links were not.

This PR corrects that