Update webservices.md with added security
Hi @pe7er,
Thanks for the shoutout, no offense to your code. You were the first to make an example and that's great for the Joomla community. We need more volunteers like you spreading the Joomla love and try contributing as much as they can at the their own pace even if it is a one character typo. Just wanted to add what I feel would be a little bit more secure and up-to-date.
For example using HTTP/2 over TLS rather than HTTP 1.1. I know there is HTTP/3 but it is not yet widely spread as HTTP/2 is.
Since it's a tutorial that potentially many people would see it would be great to use the latest and greatest as much as possible.
Thanks again my Super Joomler friend.