Update docs/combo.md
Changed opening description - it is not a comboBox like a List field as stated in original
Added hint attribute
Corrected error in example code - "/>" to close the field throws an error in some browsers. Should be "&lt;/field>"
Added note on operation from user pov to clarify not combo box
Added example modified layout to make it more comboBox like in behaviour.